+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+
+## [v7.0-beta.112] - 2021-05-06
+
+* Issue **#2223** : Fixed migration issue.
+
+
+## [v7.0-beta.111] - 2021-05-04
+
+* Change stroom dependencies to use maven central instead of bintray.
+
+* Issue **#2207** : Fixed dashboard column rename issue where column name was not updated visually after rename.
+
+
+## [v7.0-beta.110] - 2021-05-04
+
+* Issue **#2209** : Fixed more property migration issues.
+
 * Issue **#2205** : Improved migration to prevent null DB values being lost.
 
 * Issue **#2172** : Further improvements to search payload transfer and search completion.
@@ -3358,7 +3375,13 @@ Issue **gchq/stroom-expression#22** : Add `typeOf(...)` function to dashboard.
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.106...HEAD
+[Unreleased]: https://github.com/gchq/stroom/compare/v7.0-beta.112...HEAD
+[v7.0-beta.112]: https://github.com/gchq/stroom/compare/v7.0-beta.111...v7.0-beta.112
+[v7.0-beta.111]: https://github.com/gchq/stroom/compare/v7.0-beta.110...v7.0-beta.111
+[v7.0-beta.110]: https://github.com/gchq/stroom/compare/v7.0-beta.109...v7.0-beta.110
+[v7.0-beta.109]: https://github.com/gchq/stroom/compare/v7.0-beta.108...v7.0-beta.109
+[v7.0-beta.108]: https://github.com/gchq/stroom/compare/v7.0-beta.107...v7.0-beta.108
+[v7.0-beta.107]: https://github.com/gchq/stroom/compare/v7.0-beta.106...v7.0-beta.107
 [v7.0-beta.106]: https://github.com/gchq/stroom/compare/v7.0-beta.105...v7.0-beta.106
 [v7.0-beta.105]: https://github.com/gchq/stroom/compare/v7.0-beta.104...v7.0-beta.105
 [v7.0-beta.104]: https://github.com/gchq/stroom/compare/v7.0-beta.103...v7.0-beta.104